refactor: Action Type 상수 정의 및 적용
