+++ src/domains/cart/contexts/cartReducer.ts
@@ -1,4 +1,5 @@
-import { Action, State } from "../types/reducer";
+import { Action, CART_ACTION_TYPES } from "../types/cartAction";
+import { State } from "../types/cartState";
 import { CartItemWithSelection } from "../types/response";
 
 const cartReducer = (state: State, action: Action): State => {
@@ -6,13 +7,13 @@ const cartReducer = (state: State, action: Action): State => {
     items.length > 0 && items.every((item) => item.selected);
 
   switch (action.type) {
-    case "REPLACE_ITEMS":
+    case CART_ACTION_TYPES.REPLACE_ITEMS:
       return {
         items: action.items,
         allSelected: calculateAllSelected(action.items),
       };
 
-    case "TOGGLE_ALL_SELECTED": {
+    case CART_ACTION_TYPES.TOGGLE_ALL_SELECTED: {
       const newSelected = !state.allSelected;
       return {
         items: state.items.map((i) => ({ ...i, selected: newSelected })),
@@ -20,7 +21,7 @@ const cartReducer = (state: State, action: Action): State => {
       };
     }
 
-    case "TOGGLE_ITEM_SELECTED": {
+    case CART_ACTION_TYPES.TOGGLE_ITEM_SELECTED: {
       const newItems = state.items.map((i) =>
         i.id === action.id ? { ...i, selected: !i.selected } : i
       );
