+++ src/domains/cart/hooks/useCartState.ts
@@ -1,6 +1,6 @@
 import { useContext, useMemo } from "react";
 import { CartStateContext } from "../contexts/CartContext";
-import { State } from "../types/reducer";
+import { State } from "../types/cartState";
 
 export function useCartState() {
   const state = useContext(CartStateContext);
