+++ src/domains/cart/types/cartState.ts
@@ -1,0 +1,6 @@
+import { CartItemWithSelection } from "./response";
+
+export interface State {
+  items: CartItemWithSelection[];
+  allSelected: boolean;
+}
